make AuthBar a container, refactor
