+++ catalog/app/containers/AuthBar/tests/index.test.js
